+++ jest.setup.js
@@ -1,5 +1,0 @@
-const { registerAllure } = require('jest-allure2');
-
-const resultsDir = '../../../../test/results/allure';
-
-registerAllure(resultsDir);