+++ packages/tools/f-vue-icons/icons/NoteIcon.js
@@ -1,0 +1,22 @@
+import _mergeJSXProps from "babel-helper-vue-jsx-merge-props";
+export default {
+  name: 'NoteIcon',
+  props: {},
+  functional: true,
+  render: function render(h, ctx) {
+    var attrs = ctx.data.attrs || {};
+    ctx.data.attrs = attrs;
+    return h("svg", _mergeJSXProps([{
+      attrs: {
+        xmlns: "http://www.w3.org/2000/svg",
+        viewBox: "0 0 24 24"
+      },
+      "class": "c-ficon c-ficon--note"
+    }, ctx.data]), [h("path", {
+      attrs: {
+        d: "M20 11H4c-.55 0-1 .45-1 1s.45 1 1 1h16c.55 0 1-.45 1-1s-.45-1-1-1zM4 18h10c.55 0 1-.45 1-1s-.45-1-1-1H4c-.55 0-1 .45-1 1s.45 1 1 1zM20 6H4c-.55 0-1 .45-1 1v.01c0 .55.45 1 1 1h16c.55 0 1-.45 1-1V7c0-.55-.45-1-1-1z",
+        "fill-rule": "evenodd"
+      }
+    })]);
+  }
+};