+++ packages/tools/f-vue-icons/jest.config.js
@@ -25,9 +25,5 @@ module.exports = {
         'jest-serializer-vue'
     ],
 
-    testURL: 'http://localhost/',
-
-    setupFilesAfterEnv: [
-        '../../../jest.setup.js'
-    ]
+    testURL: 'http://localhost/'
 };