+++ packages/components/molecules/f-skeleton-loader/CHANGELOG.md
@@ -3,6 +3,21 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v0.3.0
+------------------------------
+*March 19, 2021*
+
+### Added
+- Skeleton components to make restaurant carousel card
+
+### Changed
+- Styles to use modules rather than scoped
+
+### Removed
+- Removed isAnimated prop
+
+
 v0.2.0
 ------------------------------
 *March 19, 2021*